requirements.txt
openai>=1.3,<2
python-dotenv>=1.0
networkx>=3.3
matplotlib>=3.8
pandas>=2.2
python-Levenshtein>=0.25
tqdm>=4.66
jinja2>=3.1
pytest>=8.1
tiktoken>=0.6
sentence-transformers==2.6.1
scikit-learn==1.5.0
